style: remove shorthand/serialize.mako.rs.
Bug: 1468651
MozReview-Commit-ID: 8Xyep2Q7trR
